==> 2_matrix_generation.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
from sklearn.preprocessing import LabelEncoder

def reduce_mem_usage(df):
    """ iterate through all the columns of a dataframe and modify the data type
        to reduce memory usage.        
    """ 
    
    for col in df.columns:
        col_type = df[col].dtype
        
        if col_type != object:
            c_min = df[col].min()
            c_max = df[col].max()
            if str(col_type)[:3] == 'int':
                if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(np.int8).max:
                    df[col] = df[col].astype(np.int8)
                elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(np.int16).max:
                    df[col] = df[col].astype(np.int16)
                elif c_min > np.iinfo(np.int32).min and c_max < np.iinfo(np.int32).max:
                    df[col] = df[col].astype(np.int32)
                elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(np.int64).max:
                    df[col] = df[col].astype(np.int64)  
            else:
                if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
                    df[col] = df[col].astype(np.float16)
                elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
                    df[col] = df[col].astype(np.float32)
                else:
                    df[col] = df[col].astype(np.float64)
        else:
            df[col] = df[col].astype('category')

    
    return df


data = []
for i in tqdm(range(1, 32)):
    data.append(reduce_mem_usage(pd.read_csv('../data/train_table_%02d.gz'%i)))
data = pd.concat(data)
data = pd.concat([data, reduce_mem_usage(pd.read_csv('../data/test_table.gz'))])

encoder = LabelEncoder()
data['id'] = encoder.fit_transform(data['link'])

data[['link','id']].to_csv('map.csv', index=False)

## 先把每天的矩阵给复原 15583 * 720
mat = np.zeros(shape=(32, 15584, 800)) - 1

idmap = pd.read_csv('map.csv')

idmap = dict(zip(idmap['link'], idmap['id']))

for date in tqdm(range(0, 31)):
    data = pd.read_csv('../data/train_table_%02d.gz'%(date+1))
    data['id'] = data['link'].apply(lambda x:idmap[x])
    for i, row in data.iterrows():
        link = int(row['id'])
        label = row['label']
        if label != -1:
            mat[date, link, int(row['predict_time'])] = label
        for loc in range(1, 6):
            mat[date, link, int(row['current_feature_time_%d'%loc])] = int(row['current_feature_status_%d'%loc])
        for w in range(1, 5):
            week = (5 - w) * 7
            if date - week >= 0:
                for loc in range(1, 6):
                    origin = mat[date - week, link, int(row['history%d_feature_time_%d'%(w, loc)])]
                    value = int(row['history%d_feature_status_%d'%(w, loc)])
                    if origin != -1 and origin != value:
                        print(origin, value, w)
                    else:
                        mat[date - week, link, int(row['history%d_feature_time_%d'%(w, loc)])] = value

date = 31
data = pd.read_csv('../data/test_table.gz')
data['id'] = data['link'].apply(lambda x:idmap[x])
count = 0
for i, row in data.iterrows():
    link = int(row['id'])
    label = row['label']
    if label != -1:
        mat[date, link, int(row['predict_time'])] = label
    for loc in range(1, 6):
        mat[date, link, int(row['current_feature_time_%d'%loc])] = int(row['current_feature_status_%d'%loc])
    for w in range(1, 5):
        week = (5 - w) * 7
        if date - week >= 0:
            for loc in range(1, 6):
                origin = mat[date - week, link, int(row['history%d_feature_time_%d'%(w, loc)])]
                value = int(row['history%d_feature_status_%d'%(w, loc)])
                if origin != -1 and origin != value:
                    count += 1
                else:
                    mat[date - week, link, int(row['history%d_feature_time_%d'%(w, loc)])] = value

np.save('mat.npy',mat)
